requirements.txt
bcrypt==4.2.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
decorator==5.1.1
Deprecated==1.2.18
fabric==3.2.2
idna==3.10
invoke==2.2.0
paramiko==3.5.1
pycparser==2.22
PyNaCl==1.5.0
pyspnego==0.11.2
pywinrm==0.5.0
requests==2.32.3
requests_ntlm==1.3.0
urllib3==2.3.0
wrapt==1.17.2
xmltodict==0.14.2
